Build serde without default features (no_std)
Our `alloc`-based compilation is based on a lack of the `"std"` feature,
but `serde` wants an explicit `"alloc"` feature enabled for `Vec`.
Instead, we can use it completely `no_std` by manually handling the
sequence of data.

Our `serde_test`-based testing is moved to a CI crate, so that doesn't
"infect" the features of the main crate.