Rename the internal unsigned_abs to avoid upstream
Rust is adding inherent `unsigned_abs` methods, which raises a forward-
compatibility warning against the use of our internal trait-method by
the same name. The inherent methods will take precedence eventually, but
that will be fine since they do the same thing. Still, the warning is
annoying, so rename ours to `uabs` just to avoid that.