Change Debug to match Display
The derived `Debug` showed the raw digit data, which is sometimes useful
if you're debugging `num-bigint` itself, but annoying in many other
contexts. For something like a simple `dbg!` call in a program, you
probably just want to see the numeric value.

This is not _necessarily_ a breaking change, but we'll treat it so.