Impl `zeroize` without using `zeroize_derive`
The `zeroize_derive` crate has quite a few dependencies, and in
particular `syn` is a fairly large one with not-insignificant compile
times.

The usages of `zeroize` are fairly trivial with two impls zeroing a
total of three fields, which makes them easy to write by hand.

It takes more crate dependencies for the proc macro to do that
(`proc-macro2`, `quote`, `syn`, `synstructure`, and `zeroize_derive`
itself) than there are fields being zeroized, so I think it makes sense
to ditch the additional dependencies and write these impls by hand.