Carry only single word in multiplication
The carry from a multiply or a multiply-accumulate is a single word,
so update the functions to clarify that. Use a temporary double word for the
computation. And rename "acc" to "carry" to distinguish from the accumulator.
Also use multiply instead of multiply-accumulate for string conversions.
And remove an unnecessary mask in get_lo().