Implement ties-to-even for Big[U]Int::to_{f32,f64}
BigUInt::to_{f32,f64} uses only the highest 64 bits as the mantissa,
discarding the rest, which gives an incorrect result when later
converted to float using IEEE-754's default nearest-ties-to-even
rounding.

Fixed to use correct round-to-odd for taking the mantissa, which means
setting the LSB to 1 if any of the lower (i.e. rounded away) bits are 1.

This also fixes the problem for BigInt since it really just uses BigUInt
under the hood.