Commit to `alloc` as much as possible
Since our MSRV is 1.60, and `alloc` was stabilized way back in 1.36, we
don't need to be conditional about using it anymore. There are still a
few places where we use `std` functionality, like `std::error::Error`,
so that feature flag is still useful, but less than it used to be.